requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Dance==6.2.0
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
py-dotenv==0.1
requests==2.28.2
requests-oauthlib==1.3.1
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
URLObject==2.4.3
Werkzeug==2.2.3
